refactor: simplify sourcemap preserver plugin
